jobs/train-model/requirements.txt
# Train Model Job Dependencies
google-cloud-aiplatform==1.*
google-cloud-storage==2.*
python-dotenv==1.*